Extract canvas helper modules
