+++ src/components/canvas/profilePrimitives.ts
@@ -1,0 +1,68 @@
+import type { Point, Segment, SketchProfile } from '../../types/project'
+import { worldToCanvas } from './viewTransform'
+import type { ViewTransform } from './viewTransform'
+
+export function anchorPointForIndex(profile: SketchProfile, index: number): Point {
+  return index === 0 ? profile.start : profile.segments[index - 1].to
+}
+
+export function arcControlPoint(start: Point, segment: Extract<Segment, { type: 'arc' }>): Point {
+  const startAngle = Math.atan2(start.y - segment.center.y, start.x - segment.center.x)
+  const endAngle = Math.atan2(segment.to.y - segment.center.y, segment.to.x - segment.center.x)
+  const radius = Math.hypot(start.x - segment.center.x, start.y - segment.center.y)
+
+  let sweep = endAngle - startAngle
+  if (segment.clockwise && sweep > 0) {
+    sweep -= Math.PI * 2
+  } else if (!segment.clockwise && sweep < 0) {
+    sweep += Math.PI * 2
+  }
+
+  const midAngle = startAngle + sweep / 2
+  return {
+    x: segment.center.x + Math.cos(midAngle) * radius,
+    y: segment.center.y + Math.sin(midAngle) * radius,
+  }
+}
+
+export function traceProfilePath(
+  ctx: CanvasRenderingContext2D,
+  profile: SketchProfile,
+  vt: ViewTransform,
+): void {
+  ctx.beginPath()
+  const start = worldToCanvas(profile.start, vt)
+  ctx.moveTo(start.cx, start.cy)
+
+  let current = profile.start
+
+  for (const segment of profile.segments) {
+    const to = worldToCanvas(segment.to, vt)
+
+    if (segment.type === 'line') {
+      ctx.lineTo(to.cx, to.cy)
+      current = segment.to
+      continue
+    }
+
+    if (segment.type === 'bezier') {
+      const control1 = worldToCanvas(segment.control1, vt)
+      const control2 = worldToCanvas(segment.control2, vt)
+      ctx.bezierCurveTo(control1.cx, control1.cy, control2.cx, control2.cy, to.cx, to.cy)
+      current = segment.to
+      continue
+    }
+
+    const center = worldToCanvas(segment.center, vt)
+    const radius = Math.hypot(current.x - segment.center.x, current.y - segment.center.y) * vt.scale
+    const startAngle = Math.atan2(current.y - segment.center.y, current.x - segment.center.x)
+    const endAngle = Math.atan2(segment.to.y - segment.center.y, segment.to.x - segment.center.x)
+
+    ctx.arc(center.cx, center.cy, radius, startAngle, endAngle, segment.clockwise)
+    current = segment.to
+  }
+
+  if (profile.closed) {
+    ctx.closePath()
+  }
+}
